CLI tests utils: bumped cli-testlab, ref migrationMatchesStub
- bumped cli-testlab to 1.7.0
- now using new fileHelper glob reader
