+++ package.json
@@ -54,7 +54,7 @@
     "JSONStream": "^1.3.5",
     "chai": "^4.2.0",
     "chai-subset-in-order": "^2.1.3",
-    "cli-testlab": "^1.6.0",
+    "cli-testlab": "^1.7.0",
     "coveralls": "^3.0.4",
     "cross-env": "^5.2.0",
     "dtslint": "^0.8.0",
