+++ test/cli/migrate-make.spec.js
@@ -1,56 +1,14 @@
 'use strict';
 
-const fs = require('fs');
 const path = require('path');
 const { FileTestHelper, execCommand } = require('cli-testlab');
 const expect = require('chai').expect;
 
 const KNEX = path.normalize(__dirname + '/../../bin/cli.js');
-
-// start: utils for --stub option tests //
-const migrationStubSetup = (fileHelper) => {
-  fileHelper.registerGlobForCleanup(
-    'test/jake-util/knexfile_migrations/*_somename.js'
-  );
-  fileHelper.createFile(
-    process.cwd() + '/knexfile.js',
-    `
-      module.exports = {
-        development: {
-          client: 'sqlite3',
-          connection: {
-            filename: __dirname + '/test/jake-util/test.sqlite3',
-          },
-          migrations: {
-            directory: __dirname + '/test/jake-util/knexfile_migrations',
-          },
-        }
-      };
-      `,
-    { isPathAbsolute: true }
-  );
-};
-
-const getMigrationFileContents = (fileHelper) => {
-  const [migrationFileName] = fs
-    .readdirSync(path.resolve(__dirname, '../jake-util/knexfile_migrations'))
-    .filter((fileName) => /\d+_somename\.js/.test(fileName));
-
-  return fileHelper.getFileTextContent(
-    `knexfile_migrations/${migrationFileName}`
-  );
-};
-
-const migrationMatchesStub = (stubPath, fileHelper) => {
-  // accepts full or relative stub path
-  const relativeStubPath = stubPath.replace('test/jake-util/', '');
-
-  const migrationContents = getMigrationFileContents(fileHelper);
-  const stubContents = fileHelper.getFileTextContent(relativeStubPath);
-  return stubContents === migrationContents;
-};
-
-// end: utils for --stub option tests //
+const {
+  migrationStubOptionSetup,
+  migrationMatchesStub,
+} = require('./cli-test-utils');
 
 describe('migrate:make', () => {
   /**
@@ -261,7 +219,7 @@ development: {
   });
 
   it('Create a new migration using --stub </file/path> relative to the knexfile', async () => {
-    migrationStubSetup(fileHelper);
+    const { migrationGlobPath } = migrationStubOptionSetup(fileHelper);
 
     const stubPath = 'test/jake-util/migration-stubs/table.stub';
 
@@ -276,11 +234,13 @@ development: {
       'test/jake-util/knexfile_migrations/*_somename.js'
     );
     expect(fileCount).to.equal(1);
-    expect(migrationMatchesStub(stubPath, fileHelper)).equal(true);
+    expect(migrationMatchesStub(stubPath, migrationGlobPath, fileHelper)).equal(
+      true
+    );
   });
 
   it('Create a new migration with --stub <name> in config.migrations.directory', async () => {
-    migrationStubSetup(fileHelper);
+    const { migrationGlobPath } = migrationStubOptionSetup(fileHelper);
 
     const stubName = 'table-foreign.stub';
     const stubPath = `test/jake-util/knexfile_migrations/${stubName}`;
@@ -296,11 +256,13 @@ development: {
       'test/jake-util/knexfile_migrations/*_somename.js'
     );
     expect(fileCount).to.equal(1);
-    expect(migrationMatchesStub(stubPath, fileHelper)).equal(true);
+    expect(migrationMatchesStub(stubPath, migrationGlobPath, fileHelper)).equal(
+      true
+    );
   });
 
   it('Create a new migration with --stub <name> when file does not exist', async () => {
-    migrationStubSetup(fileHelper);
+    migrationStubOptionSetup(fileHelper);
     const stubName = 'non-existat.stub';
     await execCommand(
       `node ${KNEX} migrate:make somename --stub ${stubName} --knexpath=../knex.js`,
@@ -311,7 +273,7 @@ development: {
   });
 
   it('Create a new migration with --stub </file/path> when file does not exist', async () => {
-    migrationStubSetup(fileHelper);
+    migrationStubOptionSetup(fileHelper);
     const stubPath = '/path/non-existat.stub';
     await execCommand(
       `node ${KNEX} migrate:make somename --stub ${stubPath} --knexpath=../knex.js`,
